Disable fail-fast in GitHub Actions
Due to the nature of robotology-superbuild, often the unstable branches build fails. However, it is important that even if the unstable branches build fails the stable branches build continue to build, so we are still able to detect regressions on the stable branches, and  just merge PRs if the stable branches pass.