Add missing dependencies of iDynTree on OsqpEigen  (#408)
* Add missing dependencies of iDynTree on OsqpEigen

Before this fix, whetever iDynTree was compiled or not with OsqpEigen support was non 
deterministic, and based on wheter OsqpEigen  was already compiled and  installed when
iDynTree was configured. With  this fix, the dependency of iDynTree on OsqpEigen is explicitly 
indicated, and the IDYNTREE_USES_* options are explicity set, so that if for some reason one
of the dependency is not installed correctly, iDynTree configuration will fail.