Remove brew cask install xquartz
The GitHub Actions macOS has been failing with:
~~~
Error: Calling brew cask install is disabled! Use brew install [--cask] instead.
Error: Process completed with exit code 1.
~~~

I do not know the root cause of this, but I hope that this command was not actually necessary as we do not have it in the README, so I think we can try to remove it.